refactor: move sidebar state to AppLayout component
